feat(helm)!: Allows for the CSI Provisioner Deployment and CSI Node Driver DaemonSet to be configured independently (#334)
* chore: Add docs to run modes

* chore(helm)!: Rename resources, labels, and containers

Note: This tries to align as well as possible to the rather scattered and vague CSI terminology.

The charts should be uninstalled and reinstalled to ensure no orphaned and conflicting resources remain.

This could impact monitoring setups that depend on the old resource names, labels, and container names.

* chore(helm)!: Use with to set context for the Deployment/DaemonSet.

BREAKING: Helm values that were under `csiNodeDriverRegistrar` are now under `csiNodeDriver`.
The reason being is that the "registrar" is only one of the containers in the DaemonSet for the Node Driver.

* chore(helm): Rename csi-provisioner deployment filename

* chore(helm)!: Separate configs for the Provisioner Deployment

* chore(helm)!: Separate configs for the NodeDriver DaemonSet

* chore(helm): Show example for overriding the operator image tag

* chore(helm): Remove "with" directive because Deployment/DaemonSet containers are not optional

NOTE: These were incorrectly introduced in 463592a while trying to hash out the structure of the values.

* feat(helm): Allow Pod priority, priorityClassName, and preemptionPolicy to be configured

* chore: Update changelog

* chore: Disable yamllint for comments-indentation

This would be rolled out by https://github.com/stackabletech/operator-templating/pull/551

* fix(helm): Remove unconfigurable Pod fields

These were introduced in 010207b. Only priorityClassName is valid.

* chore: Update changelog

* fix(helm): Rename nodeRegistrar to nodeDriverRegistrar for consistency

* Apply suggestions from code review

Co-authored-by: Techassi <git@techassi.dev>

---------

Co-authored-by: Techassi <git@techassi.dev>